Fix Pass 4 doing integer division instead of float

+++ src/main/kotlin/wotw/server/database/model/LeagueSeason.kt
@@ -278,7 +278,7 @@ class LeagueSeason(id: EntityID<Long>) : LongEntity(id) {
                     return@let
                 }
 
-                val additionalMultiplier = pointsToBoostLeft / submissionNearestToAverage.points
+                val additionalMultiplier = pointsToBoostLeft / submissionNearestToAverage.points.toDouble()
                 submissionNearestToAverage.rankingMultiplier += additionalMultiplier.toFloat()
             }
